Update and rename 二.cpp to second.cpp
